sampler: move MockEngine into testutil
