+++ sampler/coresampler_test.go
@@ -82,39 +82,3 @@ func TestAddSampleRate(t *testing.T) {
 	AddSampleRate(&root, 0.5)
 	assert.Equal(0.2, root.Metrics["_sample_rate"], "sample rate should be 20%% (50%% of 40%%)")
 }
-
-// MockEngine mocks a sampler engine
-type MockEngine struct {
-	wantSampled bool
-	wantRate    float64
-}
-
-// NewMockEngine returns a MockEngine for tests
-func NewMockEngine(wantSampled bool, wantRate float64) *MockEngine {
-	return &MockEngine{wantSampled: wantSampled, wantRate: wantRate}
-}
-
-// Sample returns a constant rate
-func (e *MockEngine) Sample(_ model.Trace, _ *model.Span, _ string) (bool, float64) {
-	return e.wantSampled, e.wantRate
-}
-
-// Run mocks Engine.Run()
-func (e *MockEngine) Run() {
-	return
-}
-
-// Stop mocks Engine.Stop()
-func (e *MockEngine) Stop() {
-	return
-}
-
-// GetState mocks Engine.GetState()
-func (e *MockEngine) GetState() interface{} {
-	return nil
-}
-
-// GetType mocks Engine.GetType()
-func (e *MockEngine) GetType() EngineType {
-	return EngineType(0)
-}
